Setup the navigation in the storyboard in a very basic sense
Setup view controller classes for the Recorder, Archive, Editor, along with the classes for the table cells within the archive.

Ended trying to figure out how the archive file system is going to work. You can have different cell classes within the same table, so that should make it easier to develop a file system within it.