The archive's navigation functions properly now.
Adding a folder will work by adding a new row to the top of the tableview that will have a textfield where a name can be entered.
I was going to use a dialogue box presented as a pop over but I could not get it to work after like 4+ hours of trying, so I'm not worrying about it.