-RecorderViewController
If we find time/it necessary, in order for the plan to save the audio files (save to a temporary url, then copy the file to another url with the name they inputted as the file) recording names can't repeat, at least within the same directory.

So we can just test against the same folder and append the folder name to the file name(in which case folder names within the same folder couldn't repeat). Again, this isn't a big issue and it can wait.